5767 - Repository: Include links and folders in ZIP (#5849)
* 5767 - Repository: Include links and folders in ZIP

- Also clean up api URLs: Remove tree in favor of many

* 5767 - Repository: getManyTree -> getMany
