+++ src/client/components/RepositoryList/Header/Header.tsx
@@ -24,7 +24,7 @@ const Header: React.FC<Props> = (props) => {
   const { isGlobal = false } = props
   const { t } = useTranslation()
   const { parentUuid, selectable } = useRepositoryListContext()
-  const path = `${ApiEndPoint.CycleData.Repository.tree()}?global=${isGlobal}`
+  const path = `${ApiEndPoint.CycleData.Repository.many()}?global=${isGlobal}`
   const fileTypeOptions = useFileTypeOptions(path)
 
   return (
