--- src/server/api/cycleData/repository/getManyRepositoryTree.ts
@@ -1,22 +1,0 @@
-import { Response } from 'express'
-
-import { CountryRequest } from 'meta/api/request/country'
-
-import { CycleDataController } from 'server/controller/cycleData'
-import Requests from 'server/utils/requests'
-
-type Request = CountryRequest<{ global: string }>
-
-export const getManyRepositoryTree = async (req: Request, res: Response): Promise<void> => {
-  try {
-    const { countryIso, global = 'false' } = req.query
-    const { assessment, cycle } = req.context
-
-    const props = { assessment, cycle, countryIso, global: JSON.parse(global) }
-    const tree = await CycleDataController.Repository.getManyTree(props)
-
-    Requests.send(res, tree)
-  } catch (e) {
-    Requests.sendErr(res, e)
-  }
-}
