+++ src/server/controller/cycleData/repository/getManyFiles.ts
@@ -1,14 +1,14 @@
 import { Readable } from 'stream'
-import { Promises } from 'utils/promises'
 
 import { CountryIso } from 'meta/area/countryIso'
 import { Assessment } from 'meta/assessment/assessment'
 import { Cycle } from 'meta/assessment/cycle'
+import { RepositoryItemTree } from 'meta/cycleData/repository/item'
+import { RepositoryItems } from 'meta/cycleData/repository/items'
 import { Lang } from 'meta/lang'
 import { Translations } from 'meta/translation/translations'
 
 import { RepositoryRepository } from 'server/db/repository/assessmentCycle/repository'
-import { FileRepository } from 'server/db/repository/public/file'
 import { FileStorage } from 'server/service/fileStorage'
 
 type Props = {
@@ -18,31 +18,36 @@ type Props = {
   global: boolean
 }
 
-type Returned = Array<{
+type FileEntry = {
   fileName: string
-  file: Readable
-}>
+  file: Buffer | Readable
+}
+
+type Returned = Array<FileEntry>
+
+const collectFileEntries = async (items: Array<RepositoryItemTree>, prefix = ''): Promise<Array<FileEntry>> => {
+  const nested = await Promise.all(
+    items.map(async (item): Promise<Array<FileEntry>> => {
+      if (RepositoryItems.isFolder(item)) {
+        return collectFileEntries(item.children, `${prefix}${item.folderName}/`)
+      }
+      const label = Translations.getLabel({ translation: item.props.translation, language: Lang.en })
+      if (item.link) {
+        const file = Buffer.from(`[InternetShortcut]\nURL=${item.link}`)
+        return [{ fileName: `${prefix}${label}.url`, file }]
+      }
+      if (item.fileUuid) {
+        const file = await FileStorage.File.get({ key: item.fileUuid })
+        return [{ fileName: `${prefix}${label}.${item.fileType}`, file }]
+      }
+      return []
+    })
+  )
+  return nested.flat()
+}
 
 export const getManyFiles = async (props: Props): Promise<Returned> => {
   const { assessment, countryIso, cycle, global } = props
-
-  const getRepositoryItemProps = { assessment, cycle, countryIso, global }
-  const repositoryItems = await RepositoryRepository.getMany(getRepositoryItemProps)
-
-  const repositoryProps = { fileUuids: repositoryItems.map((item) => item.fileUuid) }
-  const files = await FileRepository.getMany(repositoryProps)
-
-  await Promises.each(files, async (file) => {
-    const { uuid: key } = file
-    // eslint-disable-next-line no-param-reassign
-    file.file = await FileStorage.File.get({ key })
-  })
-
-  return files.map((file) => {
-    const repositoryItem = repositoryItems.find((item) => item.fileUuid === file.uuid)
-    const label = Translations.getLabel({ translation: repositoryItem.props.translation, language: Lang.en })
-    const extension = file.name.split('.').pop()
-    const fileName = `${label}.${extension}`
-    return { fileName, file: file.file }
-  })
+  const tree = await RepositoryRepository.getMany({ assessment, cycle, countryIso, global })
+  return collectFileEntries(tree)
 }
