+++ src/server/controller/cycleData/repository/index.ts
@@ -12,7 +12,6 @@ export const Repository = {
   getOneFile,
   getFileMeta,
   getMany: RepositoryRepository.getMany,
-  getManyTree: RepositoryRepository.getManyTree,
   getManyFiles,
   getOne: RepositoryRepository.getOne,
   remove,
