+++ src/server/db/repository/assessmentCycle/repository/getMany.ts
@@ -1,7 +1,8 @@
 import { AreaCode } from 'meta/area/areaCode'
 import { Assessment } from 'meta/assessment/assessment'
+import { Assessments } from 'meta/assessment/assessments'
 import { Cycle } from 'meta/assessment/cycle'
-import { RepositoryItem } from 'meta/cycleData/repository/item'
+import { RepositoryItemTree } from 'meta/cycleData/repository/item'
 import { Objects } from 'utils/objects'
 
 import { BaseProtocol, DB } from 'server/db/db'
@@ -14,24 +15,58 @@ type Props = {
   global: boolean
 }
 
-export const getMany = async (props: Props, client: BaseProtocol = DB): Promise<Array<RepositoryItem>> => {
+export const getMany = async (props: Props, client: BaseProtocol = DB): Promise<Array<RepositoryItemTree>> => {
   const { assessment, countryIso, cycle, global } = props
   const schemaCycle = Schemas.getNameCycle(assessment, cycle)
 
-  let condition = 'country_iso = $1'
+  const condition = global ? 'country_iso is null' : 'country_iso = $1'
 
-  if (global) {
-    condition = 'country_iso is null'
-  }
+  const hasODPFeature = Assessments.hasODPFeature(assessment)
+  const linkedInOdp = (ref: string): string =>
+    `exists(select 1 from ${schemaCycle}.original_data_point odp where odp.country_iso = tree.country_iso and odp.data_source_references ilike '%' || ${ref} || '%')`
+  const linkedInDescriptions = (ref: string): string =>
+    `exists(select 1 from ${schemaCycle}.descriptions d where d.country_iso = tree.country_iso and d.value::text ilike '%' || ${ref} || '%')`
 
-  return client.map<RepositoryItem>(
+  // Map of parentUuid | null : RepositoryItem
+  const nodeMap = new Map<string, RepositoryItemTree>()
+  const roots: Array<RepositoryItemTree> = []
+
+  await client.map(
     `
-      select * from ${schemaCycle}.repository
-      where
-        ${condition} and (props ->> 'hidden')::boolean is not true
-      order by id -- TODO: order by rowIndex
+      with recursive tree as (
+        select * from ${schemaCycle}.repository
+        where parent_uuid is null
+          and (props ->> 'hidden')::boolean is not true
+          and ${condition}
+
+        union all
+
+        select r.* from ${schemaCycle}.repository r
+        join tree t on r.parent_uuid = t.uuid
+        where (r.props ->> 'hidden')::boolean is not true
+      )
+      select
+        tree.*,
+        -- For items (not folders): populate "linked"
+        case when tree.folder_name is null then (
+             ${linkedInDescriptions('tree.uuid::text')}
+          or (tree.link is not null and ${linkedInDescriptions('tree.link')})
+          ${hasODPFeature ? `or ${linkedInOdp('tree.uuid::text')} or (tree.link is not null and ${linkedInOdp('tree.link')})` : ''}
+        ) end as linked,
+        -- For file items: extract extension from file.name
+        case when tree.file_uuid is not null then lower(split_part(f.name, '.', -1)) end as file_type
+      from tree
+      left join public.file f on f.uuid = tree.file_uuid
+      order by folder_name nulls last
     `,
     [countryIso],
-    (row) => Objects.camelize(row)
+    (row): void => {
+      const node: RepositoryItemTree = { ...Objects.camelize(row), children: [] }
+      nodeMap.set(node.uuid, node)
+      if (node.parentUuid) nodeMap.get(node.parentUuid)?.children.push(node)
+      else roots.push(node)
+    }
   )
+
+  return roots
 }
